+++ NEWS.md
@@ -36,6 +36,8 @@
 
 * `summary.freqlist()`: fixed a bug when all table counts are 0 and `sparse=FALSE`. (#186, #194)
 
+* `freqlist()`: implemented a `sort()` method to sort tables on frequency. (#187)
+
 # arsenal v2.0.0
 
 There is a new class system (`"arsenal_table"`) which unifies `tableby()`, `modelsum()`, and `freqlist()`.